Task09_Custom_CNN_MNIST_CIFAR10/requirements.txt
torch
torchvision
numpy
matplotlib
tqdm
streamlit
Pillow 